Make the Linux / Unix backend year 2038 compliant.

+++ src/backends/unix/system.c
@@ -84,18 +84,17 @@ Sys_Milliseconds(void)
 {
 	struct timeval tp;
 	struct timezone tzp;
-	static int secbase;
+	static long secbase;
 
 	gettimeofday(&tp, &tzp);
 
 	if (!secbase)
 	{
 		secbase = tp.tv_sec;
-		return tp.tv_usec / 1000;
+		return (int)(tp.tv_usec / 1000);
 	}
 
-	curtime = (tp.tv_sec - secbase) * 1000 + tp.tv_usec / 1000;
-
+	curtime = (int)((tp.tv_sec - secbase) * 1000 + tp.tv_usec / 1000);
 	return curtime;
 }
